feat(pan): using analog left stick
accept other filetypes mupdf uses
default fit to width
page traversal currently triangle and square
d-pad only does up and down

+++ src/filetypes/bkmudocument.h
@@ -58,8 +58,8 @@ private:
   float m_rotate;
   int m_width;
   int m_height;
-  int panX;
-	int panY;
+  float panX;
+	float panY;
 
   string filename;
